+++ src/components/Desktop/DesktopHeader.tsx
@@ -68,4 +68,6 @@ const DesktopHeader: React.FC = React.memo(() => {
   );
 });
 
+DesktopHeader.displayName = "DesktopHeader";
+
 export default DesktopHeader;
